restructure articles
