+++ apps/website/src/pages/mint-bildungsforschung.tsx
@@ -68,13 +68,13 @@ export function MintForschung({
               image: data.Dossier1TeaserImage.childImageSharp.fluid,
               title: `Wie können Mädchen fürs Programmieren begeistert werden?`,
               text: `Welche didaktischen Maßnahmen sind besonders wirksam, um das Interesse von Mädchen an Informatik, Programmieren und Computern nachhaltig zu fördern? Der folgende Artikel bietet Euch konkrete Handlungsempfehlungen aus der Forschung.`,
-              link: `/gender-didaktik-dossier_meta-review/`,
+              link: `/mint-bildungsforschung/themenspezial-artikel-maedchen-fuer-programmieren-begeistern/`,
             },
             {
               image: data.Dossier2TeaserImage.childImageSharp.fluid,
               title: `Wettbewerbe sind für Mädchen weniger attraktiv als für Jungen – über Stereotype und Strukturen`,
               text: `Mathematik- und Wissenschaftsolympiaden sind für die Teilnehmenden oft ein Sprungbrett für die Zulassung zu Spitzenuniversitäten und für künftige Karrieren im MINT-Bereich. Wie Geschlechterstereotype die Teilnahme von Mädchen beeinflussen und warum man als Anbieter:in die eigenen Wettbewerbsstrukturen kritisch reflektieren sollte, lest Ihr hier.`,
-              link: `/gender-didaktik-dossier_wettbewerbe/`,
+              link: `/mint-bildungsforschung/themenspezial-artikel-maedchen-wettbewerb/`,
             },
           ].map((teaserbox, index) => (
             <div
